Separate actions tests in different files
